+++ src/UsdnProtocol/UsdnProtocolActions.sol
@@ -42,7 +42,7 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
             timestamp, ProtocolAction.InitiateDeposit, currentPriceData
         );
 
-        _applyPnlAndFunding(currentPrice.neutralPrice, currentPrice.timestamp);
+        _applyPnlAndFunding(uint128(currentPrice.neutralPrice), uint128(currentPrice.timestamp));
         // TODO: perform liquidation of other pos with currentPrice
 
         PendingAction memory pendingAction = PendingAction({
@@ -85,7 +85,7 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
             timestamp, ProtocolAction.InitiateWithdrawal, currentPriceData
         );
 
-        _applyPnlAndFunding(currentPrice.neutralPrice, currentPrice.timestamp);
+        _applyPnlAndFunding(uint128(currentPrice.neutralPrice), uint128(currentPrice.timestamp));
         // TODO: perform liquidation of other pos with currentPrice
 
         PendingAction memory pendingAction = PendingAction({
@@ -133,14 +133,13 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
             timestamp, ProtocolAction.InitiateOpenPosition, currentPriceData
         );
 
-        // FIXME: use neutral price here!
-        _applyPnlAndFunding(currentPrice.neutralPrice, currentPrice.timestamp);
+        _applyPnlAndFunding(uint128(currentPrice.neutralPrice), uint128(currentPrice.timestamp));
 
         // TODO: perform liquidation of other pos with currentPrice
 
         // Apply liquidation penalty
         // reverts if liquidationPrice >= entryPrice
-        uint40 leverage = getLeverageWithLiquidationPenalty(currentPrice.price, liquidationPrice_);
+        uint40 leverage = getLeverageWithLiquidationPenalty(uint128(currentPrice.price), liquidationPrice_);
         if (leverage < _minLeverage) {
             revert UsdnProtocolLeverageTooLow();
         }
@@ -149,13 +148,13 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
         }
 
         // Liquidation price must be at least x% below current price
-        _checkSafetyMargin(currentPrice.price, liquidationPrice_);
+        _checkSafetyMargin(uint128(currentPrice.price), liquidationPrice_);
 
         // Register position and adjust contract state
         Position memory long = Position({
             user: msg.sender,
             amount: amount,
-            startPrice: currentPrice.price,
+            startPrice: uint128(currentPrice.price),
             leverage: leverage,
             timestamp: timestamp
         });
@@ -203,7 +202,7 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
             timestamp, ProtocolAction.InitiateClosePosition, currentPriceData
         );
 
-        _applyPnlAndFunding(currentPrice.neutralPrice, currentPrice.timestamp);
+        _applyPnlAndFunding(uint128(currentPrice.neutralPrice), uint128(currentPrice.timestamp));
         // TODO: perform liquidation of other pos with currentPrice
 
         PendingAction memory pendingAction = PendingAction({
@@ -254,12 +253,11 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
             _oracleMiddleware.parseAndValidatePrice{ value: msg.value }(deposit.timestamp, action, priceData);
 
         // adjust balances
-        // FIXME: use neutral price here!
-        _applyPnlAndFunding(depositPrice.neutralPrice, depositPrice.timestamp);
+        _applyPnlAndFunding(uint128(depositPrice.neutralPrice), uint128(depositPrice.timestamp));
 
         _balanceVault += deposit.amountOrIndex;
 
-        uint256 usdnToMint = _calcMintUsdn(deposit.amountOrIndex, depositPrice.price);
+        uint256 usdnToMint = _calcMintUsdn(deposit.amountOrIndex, uint128(depositPrice.price));
         if (initializing) {
             // we mint the minimum amount of USDN to the dead address, so that the total supply never falls to zero
             _usdn.mint(DEAD_ADDRESS, MIN_USDN_SUPPLY);
@@ -294,10 +292,9 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
         );
 
         // adjust balances
-        // FIXME: use neutral price here!
-        _applyPnlAndFunding(withdrawalPrice.neutralPrice, withdrawalPrice.timestamp);
+        _applyPnlAndFunding(uint128(withdrawalPrice.neutralPrice), uint128(withdrawalPrice.timestamp));
 
-        int256 available = _vaultAssetAvailable(withdrawalPrice.price);
+        int256 available = _vaultAssetAvailable(uint128(withdrawalPrice.price));
         if (available < 0) {
             available = 0; // clamp to zero
         }
@@ -341,15 +338,14 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
         uint128 liquidationPrice = getEffectivePriceForTick(tick);
 
         // adjust balances
-        // FIXME: use neutral price here!
-        _applyPnlAndFunding(price.neutralPrice, price.timestamp);
+        _applyPnlAndFunding(uint128(price.neutralPrice), uint128(price.timestamp));
 
         // TODO: if price <= liquidationPrice, re-calculate a liquidation price based on the leverage so that the
         // position remains solvent. Emit LiquidationPriceChanged.
 
         // Apply liquidation penalty
         // reverts if liquidationPrice >= entryPrice
-        uint40 leverage = getLeverageWithLiquidationPenalty(price.price, liquidationPrice);
+        uint40 leverage = getLeverageWithLiquidationPenalty(uint128(price.price), liquidationPrice);
         // Leverage is always greater than 1 (liquidationPrice is positive).
         // Even if it drops below _minLeverage between the initiate and validate actions, we still allow it.
         if (leverage > _maxLeverage) {
@@ -361,7 +357,7 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
         // Adjust position parameters
         Position storage pos = _longPositions[_tickHash(tick)][index];
         pos.leverage = leverage;
-        pos.startPrice = price.price;
+        pos.startPrice = uint128(price.price);
 
         emit ValidatedOpenPosition(long.user, pos, tick, index, liquidationPrice);
     }
@@ -392,8 +388,7 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
         uint128 liquidationPrice = getEffectivePriceForTick(tick);
 
         // adjust balances
-        // FIXME: use neutral price here!
-        _applyPnlAndFunding(price.neutralPrice, price.timestamp);
+        _applyPnlAndFunding(uint128(price.neutralPrice), uint128(price.timestamp));
 
         Position memory pos = getLongPosition(tick, index);
 
@@ -405,7 +400,7 @@ abstract contract UsdnProtocolActions is UsdnProtocolLong {
             available = 0;
         }
 
-        int256 value = positionValue(price.price, pos.startPrice, pos.amount, pos.leverage);
+        int256 value = positionValue(uint128(price.price), pos.startPrice, pos.amount, pos.leverage);
         if (value < 0) {
             value = 0;
         }
