+++ src/UsdnProtocol/UsdnProtocolCore.sol
@@ -94,7 +94,7 @@ abstract contract UsdnProtocolCore is IUsdnProtocolErrors, IUsdnProtocolEvents, 
         PriceInfo memory currentPrice = _oracleMiddleware.parseAndValidatePrice{ value: msg.value }(
             uint128(block.timestamp), ProtocolAction.None, priceData
         );
-        _applyPnlAndFunding(currentPrice.neutralPrice, currentPrice.timestamp);
+        _applyPnlAndFunding(uint128(currentPrice.neutralPrice), uint128(currentPrice.timestamp));
     }
 
     /* --------------------------  Internal functions --------------------------- */
